search-service/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
numpy
scikit-learn
pgvector
redis
python-logging-loki
